+++ backend/src/models/category.py
@@ -5,7 +5,7 @@ from sqlalchemy import String
 from sqlalchemy.dialects.postgresql.base import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from ..infrastructure.database import Base
+from src.infrastructure.database import Base
 
 if TYPE_CHECKING:
     from .video import Video
