+++ backend/src/models/playlist.py
@@ -6,11 +6,11 @@ from sqlalchemy import Column, DateTime, ForeignKey, Index, String, Table, Text,
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship, validates
 
-from ..infrastructure.database import Base
+from src.infrastructure.database import Base
 
 if TYPE_CHECKING:
-    from .user import User
-    from .video import Video
+    from user import User
+    from video import Video
 
 
 # --- Association table (many-to-many) ---
