+++ backend/src/models/user.py
@@ -6,21 +6,21 @@ from sqlalchemy import DateTime, ForeignKey, String, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from ..infrastructure.database import Base
+from src.infrastructure.database import Base
 
 if TYPE_CHECKING:
-    from .channel import Channel
-    from .comment_reactions import CommentReaction
-    from .comments import Comment
-    from .notification import Notification
-    from .playlist import Playlist
-    from .subscription import Subscription
-    from .user_roles import Role
-    from .user_status import UserStatus
-    from .video_reactions import VideoReaction
-    from .video_views import VideoView
-    from .watch_history import WatchHistory
-    from .watch_later import WatchLater
+    from channel import Channel
+    from comment_reactions import CommentReaction
+    from comments import Comment
+    from notification import Notification
+    from playlist import Playlist
+    from subscription import Subscription
+    from user_roles import Role
+    from user_status import UserStatus
+    from video_reactions import VideoReaction
+    from video_views import VideoView
+    from watch_history import WatchHistory
+    from watch_later import WatchLater
 
 
 class User(Base):
