+++ backend/src/models/user_roles.py
@@ -5,10 +5,10 @@ from sqlalchemy import String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from ..infrastructure import Base
+from src.infrastructure import Base
 
 if TYPE_CHECKING:
-    from .user import User
+    from user import User
 
 
 class Role(Base):
