+++ backend/src/models/video_resolutions.py
@@ -5,10 +5,10 @@ from sqlalchemy import ForeignKey, Index, Integer, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from ..infrastructure.database import Base
+from src.infrastructure.database import Base
 
 if TYPE_CHECKING:
-    from .video import Video
+    from video import Video
 
 
 class VideoResolution(Base):
