+++ backend/src/models/video_views.py
@@ -6,11 +6,11 @@ from sqlalchemy import DateTime, ForeignKey, Index, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from ..infrastructure.database import Base
+from src.infrastructure.database import Base
 
 if TYPE_CHECKING:
-    from .user import User
-    from .video import Video
+    from user import User
+    from video import Video
 
 
 class VideoView(Base):
