+++ backend/src/models/watch_history.py
@@ -5,11 +5,11 @@ from typing import TYPE_CHECKING
 from sqlalchemy import UUID, DateTime, ForeignKey, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from ..infrastructure import Base
+from src.infrastructure import Base
 
 if TYPE_CHECKING:
-    from .user import User
-    from .video import Video
+    from user import User
+    from video import Video
 
 
 class WatchHistory(Base):
