+++ backend/src/models/watch_later.py
@@ -6,11 +6,11 @@ from sqlalchemy import DateTime, ForeignKey, UniqueConstraint, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from ..infrastructure.database import Base
+from src.infrastructure.database import Base
 
 if TYPE_CHECKING:
-    from .user import User
-    from .video import Video
+    from user import User
+    from video import Video
 
 
 class WatchLater(Base):
